Use WebApi (self-hosted with Owin) to simplify simulator implementation
